task7 test2

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -10,7 +10,7 @@
 
 int print_last_digit(int digit)
 {
-	long n;
+	double n;
 
 	if (digit < 0)
 	{
